IDVA6-2091 move csrf middleware, update jest.config
