+++ jest.config.ts
@@ -33,7 +33,7 @@ const config: Config = {
     // ],
 
     // Indicates which provider should be used to instrument code for coverage
-    coverageProvider: "v8",
+    // coverageProvider: "v8",
 
     // A list of reporter names that Jest uses when writing coverage reports
     // coverageReporters: [
